refactor: extract sanitizeFtsQuery to shared storage/utils.ts
Resolves #9

The function was duplicated in messages.ts and sessions.ts with identical
logic. Any future fix to FTS query sanitization would have required
updating both files. Extracted to a single shared utility.
